add a failing test case
This fails because an enum E is stored in a single word. The high bit of
the word is 0 for the c1 case and 1 for the c2 case. The extractor needs
to project an E? (its input) to an E, and then project the E to a
TestObject. However, it only projects an E? to E and then treats the
result as if it were a TestObject.